+++ src/components/SetSelector/SetSelector.tsx
@@ -1,9 +1,9 @@
 'use client';
 
-import { CardSet, fetchSets } from "@/types/CardSet";
-import { FormControl, FormLabel, MenuItem, Select } from "@mui/material";
+import { CardSet, fetchSets, setDateSort } from "@/types/CardSet";
+import { Autocomplete, FormControl, FormLabel, MenuItem, Select, TextField } from "@mui/material";
 import Image from "next/image";
-import { useEffect, useState } from "react";
+import { useEffect, useMemo, useState } from "react";
 
 type SetSelectorProps = {
     onSetChange: (set: CardSet | null) => void;
@@ -25,14 +25,33 @@ export function SetSelector({ onSetChange }: SetSelectorProps) {
         loadSets();
     }, []);
 
+    const setOptions = useMemo(() => {
+        return sets.sort(setDateSort).map(s => ({
+            label: `${s.name} (${s.releaseDate.getFullYear()})`,
+            value: s.code,
+        }));
+    }, [sets]);
+
     return (
         <FormControl sx={{ m: 3 }}>
             <FormLabel>Filter by set</FormLabel>
-            <Select value={set?.code || ''} onChange={(event) => setSet(sets.find(s => s.code === event.target.value) || null)}>
-                {sets.map(({ code, name, iconSvgUri, releaseDate }, index) => <MenuItem key={index} value={code}>
-                    <Image src={iconSvgUri} alt={code} height="16" width="16" style={{ height: '1em', verticalAlign: 'middle', marginRight: '0.5em' }} /> {name} ({releaseDate.getFullYear()})
-                </MenuItem>)}
-            </Select>
+            <Autocomplete
+                options={setOptions}
+                getOptionLabel={(option) => option.label}
+                onChange={(event, newValue) => {
+                    setSet(newValue ? sets.find(s => s.code === newValue.value) || null : null);
+                }}
+                renderInput={(params) => <TextField {...params} variant="outlined" />}
+            />
         </FormControl>
     );
 }
+//                 options={countries}
+//             <Select value={set?.code || ''} onChange={(event) => setSet(sets.find(s => s.code === event.target.value) || null)}>
+//                 {sets.map(({ code, name, iconSvgUri, releaseDate }, index) => <MenuItem key={index} value={code}>
+//                     <Image src={iconSvgUri} alt={code} height="16" width="16" style={{ height: '1em', verticalAlign: 'middle', marginRight: '0.5em' }} /> {name} ({releaseDate.getFullYear()})
+//                 </MenuItem>)}
+//             </Select>
+//         </FormControl>
+//     );
+// }